requirements.txt
dvc[gdrive]
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
Pillow>=8.3.0
opencv-python>=4.5.0
scikit-learn>=0.24.0
imagehash>=4.2.0
tqdm>=4.61.0
pyyaml>=6.0
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
ultralytics==8.3.160
ultralytics-thop==2.0.12